+++ test/tls_certificate_check_test_utils.erl
@@ -37,7 +37,8 @@ connect(PemsPath, AuthoritiesFilename, ChainOrLeaf, ChainOrLeafFilename, Fun) ->
 connect(PemsPath, AuthoritiesFilename, ChainOrLeaf, ChainOrLeafFilename, KeyFilename, Fun) ->
     AuthoritiesPath = filename:join([PemsPath, "CA_stores", AuthoritiesFilename]),
     {ok, EncodedAuthorities} = file:read_file(AuthoritiesPath),
-    ok = tls_certificate_check_shared_state:maybe_update_shared_state(EncodedAuthorities),
+    ok = tls_certificate_check_shared_state:maybe_update_shared_state(EncodedAuthorities,
+                                                                      [force_hardcoded]),
 
     {ListenSocket, Port, AcceptorPid} = start_server_with_chain(PemsPath, ChainOrLeaf,
                                                                 ChainOrLeafFilename,