+++ CHANGELOG.md
@@ -5,6 +5,16 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
+## [Unreleased]
+
+### Added
+
+- Windows to CI
+
+### Changed
+
+- default CAs to the ones trusted by OTP (typically provided by the OS), when available, on OTP 25+
+
 ## [1.16.0] - 2022-10-11
 
 ### Added